Modify to support number of total chromosomes > 65535

+++ 2bwt-flex/OCC.h
@@ -35,7 +35,7 @@
 
 typedef struct OCCPositionCache
 {
-	unsigned short ChromId;
+	unsigned int ChromId;
 	unsigned char ReadStrand;
 	unsigned long long tp;
 	int occMismatch;
